Upgraded to latest veranda version (v1.1.0). Moved command line interface to base module. Removed unused code.
