+++ setup.cfg
@@ -30,8 +30,9 @@ setup_requires = pyscaffold>=3.3a0,<4
 # Add here dependencies of your project (semicolon/line-separated), e.g.
 
 install_requires =
-    veranda==0.1.0
+    veranda
     pandas
+    numpy
 
 # The usage of test_requires is discouraged, see `Dependency Management` docs
 # tests_require = pytest; pytest-cov
