+++ src/abcraster/base.py
@@ -1,8 +1,9 @@
 import os
+import argparse
 from osgeo import ogr
 import numpy as np
 import pandas as pd
-from veranda.io.geotiff import GeoTiffFile
+from veranda.raster.native.geotiff import GeoTiffFile
 from abcraster.input import rasterize
 from abcraster.sampling import gen_random_sample
 from abcraster.metrics import metrics
@@ -40,19 +41,19 @@ class Validation:
         """
 
         # load classification result
-        with GeoTiffFile(ras_data_filepath, auto_decode=False) as src:
-            self.input_data = src.read(return_tags=False)
-            self.gt = src.geotransform
-            self.sref = src.spatialref
+        with GeoTiffFile(ras_data_filepath) as src:
+            self.input_data = src.read()[1]
+            self.gt = src.geotrans
+            self.sref = src.sref_wkt
 
         # load exclusion mask
         if ex_filepath is None:
             self.ex_mask = None
         else:
-            with GeoTiffFile(ex_filepath, auto_decode=False) as src:
-                self.ex_mask = src.read(return_tags=False)
-                ex_gt = src.geotransform
-                ex_sref = src.spatialref
+            with GeoTiffFile(ex_filepath) as src:
+                self.ex_mask = src.read()[1]
+                ex_gt = src.geotrans
+                ex_sref = src.sref_wkt
 
                 if ex_gt != self.gt or ex_sref != self.sref:
                     print("Exclusion WARNING:Grid/projection of input and reference data are not the same!")
@@ -71,13 +72,14 @@ class Validation:
                 os.remove(v_rasterized_path)
                 delete_shapefile(v_reprojected_path)
         elif ref_file_ext == '.tif':
-            with GeoTiffFile(ref_data_filepath, auto_decode=False) as src:
-                self.ref_data = src.read(return_tags=False)
-                ref_gt = src.geotransform
-                ref_sref = src.spatialref
+            with GeoTiffFile(ref_data_filepath) as src:
+                self.ref_data = src.read()[1]
+                ref_gt = src.geotrans
+                ref_sref = src.sref_wkt
 
                 if ref_gt != self.gt or ref_sref != self.sref:
                     print("WARNING: Grid/projection of input and reference data are not the same!")
+                    # TODO: transform reference data to match CRS of input data
         else:
             raise ValueError("Input file with extension " + ref_file_ext + " is not supported.")
 
@@ -120,7 +122,7 @@ class Validation:
             fp = np.sum(res == 3)
         self.confusion_matrix = np.array([[tp, fp], [fn, tn]])
 
-    def define_sampling(self, sampling, opt_stratification=False, samples_filepath=None):
+    def define_sampling(self, sampling, samples_filepath=None):
         """
         Defines samples for the validation.
 
@@ -131,8 +133,6 @@ class Validation:
             non-stratified sampling = integer number of class-independent samples
             None = this implies samples are loaded from samples_filepath (default: None),
             *sampling is superseded if samples_filepath is None
-        opt_stratification: bool, optional
-            Should a stratification be applied (default: False)?
         samples_filepath: str, optional
             Path of the output samples file, which is writen if the path is passed (default: None).
         """
@@ -142,14 +142,15 @@ class Validation:
 
         # write output
         if samples_filepath is not None:
-            with GeoTiffFile(samples_filepath, mode='w', count=1, geotransform=self.gt, spatialref=self.sref) as src:
-                src.write(self.samples, band=1, nodata=[255])
+            with GeoTiffFile(samples_filepath, mode='w', geotrans=self.gt, sref_wkt=self.sref, nodatavals=[255],
+                             overwrite=True) as src:
+                src.write(self.samples)
 
     def load_sampling(self, samples_filepath):
         """Loads the samples from a raster file."""
 
-        with GeoTiffFile(samples_filepath, auto_decode=False) as src:
-            self.samples = src.read(return_tags=False)
+        with GeoTiffFile(samples_filepath) as src:
+            self.samples = src.read()[1]
 
     def calculate_accuracy_metric(self, metric_func):
         """
@@ -157,7 +158,7 @@ class Validation:
 
         Parameters
         ----------
-        metric_func: function
+        metric_func: callable
             Function, which takes in a dictionary with the keys TP (true positives), FP (false positives), TN (true
             negatives) and FN (false negatives), and returns an accuracy metric.
 
@@ -188,8 +189,9 @@ class Validation:
         """
 
         valid = np.logical_and(self.ref_data != 255, self.input_data != 255)
-        with GeoTiffFile(valid_filepath, mode='w', count=1, geotransform=self.gt, spatialref=self.sref) as src:
-            src.write(valid, band=1, nodata=[255])
+        with GeoTiffFile(valid_filepath, mode='w', geotrans=self.gt, sref_wkt=self.sref, nodatavals=[255],
+                         overwrite=True) as src:
+            src.write(valid)
 
     def write_confusion_map(self, out_filepath):
         """
@@ -201,8 +203,9 @@ class Validation:
             Path of the output file.
         """
 
-        with GeoTiffFile(out_filepath, mode='w', count=1, geotransform=self.gt, spatialref=self.sref) as src:
-            src.write(self.confusion_map, band=1, nodata=[255])
+        with GeoTiffFile(out_filepath, mode='w', geotrans=self.gt, sref_wkt=self.sref, nodatavals=[255],
+                         overwrite=True) as src:
+            src.write(self.confusion_map)
 
 
 def delete_shapefile(shp_path):
@@ -293,7 +296,7 @@ def run(ras_data_filepaths, ref_data_filepath, out_dirpath, metrics_list, sample
 
         result = [input_base_filename, ref_base_filename]
 
-        #computes the selected metrics
+        # computes the selected metrics
         for m in metrics_list:
             metric = metrics[m]
             result += [v.calculate_accuracy_metric(metric)]
@@ -304,12 +307,12 @@ def run(ras_data_filepaths, ref_data_filepath, out_dirpath, metrics_list, sample
             v_rasterized_path = os.path.join(out_dirpath, v_rasterized_filename)
             ref_data_filepath_current = v_rasterized_path  # use the temporary rasterized reference
 
-        if i == num_inputs - 1: #  if last run use delete flag from cli
+        if i == num_inputs - 1:  # if last run use delete flag from cli
             if delete_tmp_files:
                 os.remove(os.path.join(out_dirpath, v_rasterized_filename))
                 delete_shapefile(os.path.join(out_dirpath, v_reprojected_filename))
 
-    if num_inputs == 1: #changed output format for multiple results
+    if num_inputs == 1:  # changed output format for multiple results
         df = pd.DataFrame(results[0], cols)
         print(df)
     else:
@@ -320,3 +323,101 @@ def run(ras_data_filepaths, ref_data_filepath, out_dirpath, metrics_list, sample
     return df
 
 
+def command_line_interface():
+    """ Command line interface to perform a validation based on vector-based reference data. """
+
+    # define parser
+    parser = argparse.ArgumentParser(
+        description="Simple Binary Validation Workflow. Initially designed to validate floods")
+    parser.add_argument("-in", "--input_filepaths",
+                        help="Full file path to the binary raster data 1= presence, 0=absence, for now 255=nodata.",
+                        required=True, nargs="+", type=str)
+    parser.add_argument("-ref", "--reference_filepath",
+                        help="Full file path to the validation dataset (.tif or .shp, in any projection)",
+                        required=True, type=str)
+    parser.add_argument("-ex", "--exclusion_filepath",
+                        help="Full file path to the binary exclusion data 1= exclude, for now 255=nodata.",
+                        required=False, type=str)
+    parser.add_argument("-ns", "--num_samples",
+                        help="number of total samples if sampling will be applied.", required=False, type=int)
+    parser.add_argument("-stf", "--stratify", help="Stratification based on reference data", required=False,
+                        default=True, action="store_true")
+    parser.add_argument("-nst", "--no_stratify", dest='stratify', action='store_false',
+                        help="No Stratification option")
+    parser.add_argument("-sfp", "--samples_filepath",
+                        help="Full file path to the sampling raster dataset (.tif ), if num samples not specified, \
+                        assumes samples will be read from this path",
+                        required=False, type=str)
+    parser.add_argument("-out", "--output_raster",
+                        help="Full file path to the final difference raster", required=True, type=str)
+    parser.add_argument("-csv", "--output_csv",
+                        help="Full file path to the csv results", required=False, type=str)
+    parser.add_argument("-del", "--delete_tmp",
+                        help="Option to delete temporary files.", required=False, type=bool)
+    parser.add_argument("-all", "--all_metrics", help="Option to compute all metrics.",
+                        default=True, action="store_true")
+    parser.add_argument('-na', "--not_all_metrics", dest='all_metrics', action='store_false',
+                        help="Option not to compute all metrics, metrics should be specified if activated.")
+    parser.add_argument("-mts", "--metrics", nargs="+", required=False, type=str,
+                        help="Option to list metrics to run.")
+
+    # collect inputs
+    args = parser.parse_args()
+    input_raster_filepaths = args.input_filepaths
+    exclusion_filepath = args.exclusion_filepath
+    validation_filepath = args.reference_filepath
+    output_raster_filepath = args.output_raster
+    output_csv_filepath = args.output_csv
+    delete_tmp = args.delete_tmp
+    strat = args.stratify
+    sampling = args.num_samples
+    samples_filepath = args.samples_filepath
+
+    if args.all_metrics:
+        metrics_list = metrics.keys()  # all metrics as defined in metrics dictionary
+    else:
+        metrics_list = args.metrics
+
+        if len(metrics_list) == 0:
+            raise RuntimeError('Metric keys list not found!')
+
+        for m in metrics_list:
+            if m not in metrics:
+                raise RuntimeError('Metric key ({}) not found!'.format(m))
+
+    if len(input_raster_filepaths) > 1 and output_csv_filepath is None:
+        raise RuntimeError('Output CSV file is required for multiple input files')
+
+    if args.metrics is not None and len(args.metrics) > 0 and args.all_metrics:
+        print("WARNING: Specific metrics specified but all metrics selected. Proceeding with all metrics")
+
+    if sampling is not None:
+        if strat:
+            sampling = [sampling]
+        else:
+            sampling = int(sampling)
+
+        if samples_filepath is None:
+            print("WARNING: Number of samples specified but no filepath to output samples raster specified! \
+            Ignoring number of samples. Proceeding without sampling.")  # warning or run time error?
+            sampling = None
+
+    # define output names
+    out_dirpath, out_raster_filename = os.path.split(output_raster_filepath)
+    base = os.path.splitext(out_raster_filename)[0]
+    reproj_shp_filepath = base + '_reproj_input_vector.shp'
+    rasterized_shp_filepath = base + '_rasterize_input_vector.tif'
+
+    # set default option
+    if delete_tmp is None:
+        delete_tmp = False
+
+    run(ras_data_filepaths=input_raster_filepaths, ref_data_filepath=validation_filepath, out_dirpath=out_dirpath,
+        diff_ras_out_filename=out_raster_filename, v_reprojected_filename=reproj_shp_filepath,
+        v_rasterized_filename=rasterized_shp_filepath, out_csv_filename=output_csv_filepath,
+        ex_filepath=exclusion_filepath, delete_tmp_files=delete_tmp,
+        sampling=sampling, samples_filepath=samples_filepath, metrics_list=metrics_list)
+
+
+if __name__ == '__main__':
+    command_line_interface()
