+++ src/abcraster/input.py
@@ -3,16 +3,6 @@ import numpy as np
 import os
 
 
-def get_equi7grid_geotags(tile, sres=20, continent='EU'):
-    """ Retrieve spatial details of an Equi7grid tile. """
-    from equi7grid.equi7grid import Equi7Grid  # moved import here to make equi7grid optional
-    grid = Equi7Grid(sres).subgrids[continent]
-    tile_geotags = grid.tilesys.create_tile(name=tile).get_geotags()
-    gt, sref = tile_geotags['geotransform'], tile_geotags['spatialreference']
-
-    return gt, sref
-
-
 def reproject_vec(layer, out_sref, v_reprojected_filepath='tmp.shp'):
     """
     Reprojects a vector layer to a different projection.
